1. Moved APIs.json to /docs | 2. Moved docView and renamed to /api and docs | 3. Updated Dev and Prod URLS. | 4. Updated site URL.
