+++ Swag/Info.json
@@ -5,7 +5,7 @@
     "description": "This documentation contains all information for the backend API of the Coding Garden Community App.",
     "contact": {
       "name": "The CodingGarden Community",
-      "url": "https://github.com/CodingGardenCommunity"
+      "url": "https://coding.garden/"
     },
     "license": {
       "name": "The MIT License",
