+++ package.json
@@ -11,7 +11,7 @@
     "seed": "node src/tasks/seed/localDevSeed.js",
     "deploy": "./deploy.sh",
     "format": "prettier --write src/**/*.js",
-    "buildAPIDoc": "swagger-ui-watcher ./Swag/MainSwag.json --bundle=./APIs.json"
+    "buildAPIDoc": "swagger-ui-watcher ./Swag/MainSwag.json --bundle=./docs/APIs.json"
   },
   "repository": "https://github.com/CodingGardenCommunity/app-backend.git",
   "author": "Coding Garden Community App",
