+++ src/api/docs/doc.controller.js
@@ -1,0 +1,7 @@
+const path = require('path');
+
+function sendVersionMarkup(req, res) {
+  res.sendFile(path.join(__dirname, '/versions.html'));
+}
+
+module.exports = { sendVersionMarkup };
